+++ src/cl.gallery/CHANGELOG.rst
@@ -2,6 +2,19 @@
 Cl.Gallery
 ==========
 
+2.0.0
+-----
+- rewrite gallery to support different engines (animation types)
+- reqrite internal api
+- added option ``engine``
+- added engine ``fade`` and ``slide``
+- added option ``autoplay``
+- added queue function
+- added WAI-ARIA labels
+- changed index from 0 to null
+- changed ``direction`` to ``move``
+- changed ``move`` default values to next, previous or random
+
 1.0.1
 -----
 - added option ``easing``